// wellsone-admin/WellDone/src/app/models/user.ts
export class User {
  _id: string;
  username: string;
  email: string;
  password: string;
  first_name: string;
  last_name: string;
  title: string;
  avatar: string;
  status_active: string;
  status_logged: string;
  num_of_articles: string;
  num_of_likes: string;
  history: string;
  created_at: string;
  modified_at: string;
  premium: boolean;
}